+++ src/javascripts/views/startApp.js
@@ -3,8 +3,6 @@ import logoutButton from '../components/logoutButton';
 import navbar from '../components/navbar';
 import { domEvents } from '../events/domEvents';
 import navEvents from '../events/navEvents';
-import { menuWithReservation } from '../helpers/data/menuReservationData';
-import { getSeatingReservations } from '../helpers/data/seatingReservationsData';
 import landingPage from './landingPage';
 
 const startApp = (user) => {
@@ -14,8 +12,6 @@ const startApp = (user) => {
   landingPage();
   domEvents(user);
   navEvents(user);
-  getSeatingReservations();
-  menuWithReservation();
 };
 
 export default startApp;
